IntegratedSystemTest includes smartswitches. Also, a problem with some displays is addresses
The display issue is this: Writing a long textstring (more than 16
chars) normally just continues on the second line of the display. But I
have encountered some displays where this didn't happen. Therefore: For
the future it should never be assumed that textlines are wrapped over
to the second line - rather writing to the second line should be
explicitly written.